requirements.txt
flask==3.0.3
tensorflow==2.20.0
opencv-python-headless==4.12.0.88
numpy==2.2.6
pillow==11.3.0
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
gunicorn==23.0.0 